backend/notification-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
sib-api-v3-sdk==7.5.0
twilio==8.10.0
sqlalchemy==2.0.23
python-multipart==0.0.6
jinja2==3.1.2
pytest==7.4.3